backend/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
cssselect2==0.7.0
defusedxml==0.7.1
Django==4.2.2
django-filter==23.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-yasg==1.21.5
flake8==6.0.0
html5lib==1.1
idna==3.4
importlib-metadata==6.6.0
inflection==0.5.1
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.3
mccabe==0.7.0
oauthlib==3.2.2
oscrypto==1.3.0
packaging==23.1
Pillow==9.5.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
pyHanko==0.19.0
pyhanko-certvalidator==0.23.0
PyJWT==2.7.0
pypdf==3.11.1
pypng==0.20220715.0
python-bidi==0.4.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.13
requests==2.31.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==5.0.1
uritemplate==4.1.1
uritools==4.0.1
urllib3==2.0.3
webencodings==0.5.1
xhtml2pdf==0.2.11
zipp==3.15.0
